requirements.txt
# Zero-Knowledge Password Manager Dependencies
# Install with: pip install -r requirements.txt

# Core cryptographic library
cryptography>=41.0.0

# Argon2 password hashing (memory-hard function)
argon2-cffi>=23.1.0

# Web interface framework
flask>=2.3.0
werkzeug>=2.3.0

# Development and testing dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# mypy>=1.5.0
